Test13 works

+++ request.c
@@ -34,7 +34,7 @@ void requestError(int fd, char *cause, char *errnum, char *shortmsg, char *longm
     Rio_writen(fd, buf, strlen(buf));
     printf("%s", buf);
     
-    sprintf(buf, "Content-Length: %d\r\n\r\n", strlen(body));
+    sprintf(buf, "Content-Length: %d\r\n\r\n", (int)strlen(body));
     Rio_writen(fd, buf, strlen(buf));
     printf("%s", buf);
